@@ -6,6 +6,6 @@
 //= require cloudinary/jquery.cloudinary
 //= require bootstrap
 //= require flash
-
+//= require regex
 
 
