@@ -28,8 +28,8 @@ class ApplicationController < ActionController::Base
   end
 
   def after_sign_in_path_for(resource_or_scope)
-    if current_user.validate == 1
-      current_user
+    if current_user.validation == 1
+      "/profiles"
     else
       new_user_profile_path(current_user)
     end
